% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_accordion.R
\name{mod_accordion_server}
\alias{mod_accordion_server}
\title{mod_accordion_server
Backend for accordion component}
\usage{
mod_accordion_server(id, session_db)
}
\arguments{
\item{id}{a unique ID of a specific instance of the accordion component}

\item{session_db}{an R6 object for logging data to the database}
}
\description{
mod_accordion_server
Backend for accordion component
}
