% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_accordion.R
\name{mod_accordion_ui}
\alias{mod_accordion_ui}
\title{////////////////////////////////////////////////////////////////////////////
FILE: mod_accordion.R
AUTHOR: David Ruvolo
CREATED: 2020-07-03
MODIFIED: 2020-07-03
PURPOSE: accordion input
STATUS: in.progress
PACKAGES: Shiny
COMMENTS: See custom input binding
////////////////////////////////////////////////////////////////////////////
mod_accordion_ui
Create an accordion element that allows you to open/close a "hidden"
section}
\usage{
mod_accordion_ui(id, title, content)
}
\arguments{
\item{id}{a unique id for this instance of the accordion component}

\item{title}{a title for the accordion component}

\item{content}{a Shiny Tag element or tagList}
}
\description{
////////////////////////////////////////////////////////////////////////////
FILE: mod_accordion.R
AUTHOR: David Ruvolo
CREATED: 2020-07-03
MODIFIED: 2020-07-03
PURPOSE: accordion input
STATUS: in.progress
PACKAGES: Shiny
COMMENTS: See custom input binding
////////////////////////////////////////////////////////////////////////////
mod_accordion_ui
Create an accordion element that allows you to open/close a "hidden"
section
}
