% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_login.R
\name{mod_login_server}
\alias{mod_login_server}
\title{login Server Function}
\usage{
mod_login_server(id, data, logged, session_db)
}
\arguments{
\item{id}{unique ID for an instance of a module}

\item{data}{an object containing the user accounts}

\item{logged}{a reactive object that manages the app logged in state}

\item{session_db}{a R6 object used for sending data to the database}
}
\description{
login Server Function
}
