% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_login.R
\name{mod_login_ui}
\alias{mod_login_ui}
\title{////////////////////////////////////////////////////////////////////////////
FILE: mod_login.R
AUTHOR: David Ruvolo
CREATED: 2020-06-27
MODIFIED: 2020-07-14
PURPOSE: module for login page
STATUS: working
PACKAGES: NA
COMMENTS:
This script provides two modules: the login UI and the login Server
The UI component renders form with the two input elements for username and
password. A back to main page link is also provided.}
\usage{
mod_login_ui(id, class)
}
\arguments{
\item{id}{a unique ID for this instance of the login UI}

\item{class}{a css class to include in the rendering of the of UI
//////////////////////////////////////
Username: label, input, error
//////////////////////////////////////
Password: label, input, error}
}
\description{
The server logic runs only when the submit button is clicked. When clicked,
a series of server-side form validation is executed. If there are any
errors, a message is sent to the client and the error is added to the
database. When all errors are eliminated and the user has provided valid
credentials, then the app state is changed (i.e., logged in) and the state
is noted in the database.
////////////////////////////////////////////////////////////////////////////
ui component
}
