% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_navigation.R
\name{mod_navigation_ui}
\alias{mod_navigation_ui}
\title{mod_navigation_ui}
\usage{
mod_navigation_ui(id, buttons)
}
\arguments{
\item{id}{a unique ID for the component}

\item{buttons}{the buttons you like to render (previous, next, submit, etc.)}
}
\description{
Create a navigation component by naming buttons
}
