% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_side_effect_card.R
\name{mod_se_server}
\alias{mod_se_server}
\title{mod_se_server}
\usage{
mod_se_server(id, session_db)
}
\arguments{
\item{id}{unique ID for this module}

\item{session_db}{an R6 object used for sending data to the db
session_db$capture_action(
    event = "side_effect_selection",
    id = id,
    desc = "user deselected side effect"
)}
}
\description{
mod_se_server
}
