% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_side_effect_card.R
\name{mod_side_effect_ui}
\alias{mod_side_effect_ui}
\title{////////////////////////////////////////////////////////////////////////////
FILE: mod_side_effect_card.R
AUTHOR: David Ruvolo
CREATED: 2020-06-27
MODIFIED: 2020-07-03
PURPOSE: side effects card component + server module
STATUS: in.progress
PACKAGES: NA
COMMENTS: NA
////////////////////////////////////////////////////////////////////////////
mode_side_effect_ui
Create an accordion component with collapsible content}
\usage{
mod_side_effect_ui(id, title, text)
}
\arguments{
\item{id}{a unique ID for the component}

\item{title}{a title for the card}

\item{text}{content to be placed inside the collapsible section}
}
\description{
////////////////////////////////////////////////////////////////////////////
FILE: mod_side_effect_card.R
AUTHOR: David Ruvolo
CREATED: 2020-06-27
MODIFIED: 2020-07-03
PURPOSE: side effects card component + server module
STATUS: in.progress
PACKAGES: NA
COMMENTS: NA
////////////////////////////////////////////////////////////////////////////
mode_side_effect_ui
Create an accordion component with collapsible content
}
